Fixed paged list type bug.

+++ RecipesManagerApi.Application/Paging/PagedList.cs
@@ -16,7 +16,7 @@
 
         public PagedList() { }
 
-        public PagedList(IEnumerable<T> items, PageParameters pageParameters, long totalItems)
+        public PagedList(IEnumerable<T> items, PageParameters pageParameters, int totalItems)
         {
             this.PageNumber = pageParameters.PageNumber;
             this.PageSize = pageParameters.PageSize;
